Removed smooth, minmax, where and convol from datatypes.[c|h]pp to let them live their own life. This improves an awful lot the speed of compilation on multicores and reduce the individual memory imprint.
