+++ src/basic_fun.hpp
@@ -95,7 +95,6 @@ namespace lib {
   BaseGDL* strpos( EnvT* e);
   BaseGDL* strtrim( EnvT* e);
 
-  BaseGDL* where_fun( EnvT* e);
   BaseGDL* total_fun( EnvT* e);
   BaseGDL* product_fun( EnvT* e);
