+++ src/datatypes.cpp
@@ -4017,939 +4017,6 @@ bool Data_<SpDObj>::LogTrue(SizeT i)
 {
   return (*this)[i] != 0;
 }
-// structs are not allowed
-template<class Sp>
-DByte* Data_<Sp>::TagWhere(SizeT& n)
-{
-  SizeT nEl = N_Elements();
-  DByte* ixList = new DByte[ nEl];
-  SizeT count = 0;
-#pragma omp parallel reduction(+:count) if (nEl >= CpuTPOOL_MIN_ELTS && (CpuTPOOL_MAX_ELTS == 0 || CpuTPOOL_MAX_ELTS <= nEl))
-    {
-#pragma omp for 
-    for (SizeT i = 0; i < nEl; ++i) {
-      DByte tmp=(dd[i]!=0);
-      ixList[i]=tmp;
-      count +=tmp;
-    }
-    }
-  n = count;
-  return ixList;    
-}
-
-template<>
-DByte* Data_<SpDFloat>::TagWhere(SizeT& n) {
-  SizeT nEl = N_Elements();
-  DByte* ixList = new DByte[ nEl];
-  SizeT count = 0;
-#pragma omp parallel reduction(+:count) if (nEl >= CpuTPOOL_MIN_ELTS && (CpuTPOOL_MAX_ELTS == 0 || CpuTPOOL_MAX_ELTS <= nEl))
-  {
-#pragma omp for 
-    for (SizeT i = 0; i < nEl; ++i) {
-      DByte tmp = (dd[i] != 0.0f);
-      ixList[i] = tmp;
-      count += tmp;
-    }
-  }
-  n = count;
-  return ixList;
-}
-
-template<>
-DByte* Data_<SpDDouble>::TagWhere(SizeT& n) {
-  SizeT nEl = N_Elements();
-  DByte* ixList = new DByte[ nEl];
-  SizeT count = 0;
-#pragma omp parallel reduction(+:count) if (nEl >= CpuTPOOL_MIN_ELTS && (CpuTPOOL_MAX_ELTS == 0 || CpuTPOOL_MAX_ELTS <= nEl))
-  {
-#pragma omp for 
-    for (SizeT i = 0; i < nEl; ++i) {
-      DByte tmp = (dd[i] != 0.0);
-      ixList[i] = tmp;
-      count += tmp;
-    }
-  }
-  n = count;
-  return ixList;
-}
-template<>
-DByte* Data_<SpDString>::TagWhere(SizeT& n)
-{
-  SizeT nEl = N_Elements();
-  DByte* ixList = new DByte[ nEl];
-  SizeT count = 0;
-#pragma omp parallel reduction(+:count) if (nEl >= CpuTPOOL_MIN_ELTS && (CpuTPOOL_MAX_ELTS == 0 || CpuTPOOL_MAX_ELTS <= nEl))
-    {
-#pragma omp for 
-    for (SizeT i = 0; i < nEl; ++i) {
-      DByte tmp=((*this)[i]!="");
-      ixList[i]=tmp;
-      count +=tmp;
-    }
-    }
-  n = count;
-  return ixList;    
-}
-template<>
-DByte* Data_<SpDComplex>::TagWhere(SizeT& n)
-{
-  SizeT nEl = N_Elements();
-  DByte* ixList = new DByte[ nEl];
-  SizeT count = 0;
-#pragma omp parallel reduction(+:count) if (nEl >= CpuTPOOL_MIN_ELTS && (CpuTPOOL_MAX_ELTS == 0 || CpuTPOOL_MAX_ELTS <= nEl))
-    {
-#pragma omp for 
-    for (SizeT i = 0; i < nEl; ++i) {
-      DFloat re=(*this)[i].real();
-      DFloat im=(*this)[i].imag();
-      DByte tmp=(re!=0.0f || im!=0.0f);
-      ixList[i]=tmp;
-      count +=tmp;
-    }
-    }
-  n = count;
-  return ixList;    
-}
-template<>
-DByte* Data_<SpDComplexDbl>::TagWhere(SizeT& n)
-{
-  SizeT nEl = N_Elements();
-  DByte* ixList = new DByte[ nEl];
-  SizeT count = 0;
-#pragma omp parallel reduction(+:count) if (nEl >= CpuTPOOL_MIN_ELTS && (CpuTPOOL_MAX_ELTS == 0 || CpuTPOOL_MAX_ELTS <= nEl))
-    {
-#pragma omp for 
-    for (SizeT i = 0; i < nEl; ++i) {
-      DDouble re=(*this)[i].real();
-      DDouble im=(*this)[i].imag();
-      DByte tmp=(re!=0.0 || im!=0.0);
-      ixList[i]=tmp;
-      count +=tmp;
-    }
-    }
-  n = count;
-  return ixList;    
-}
-
-// AC: see bug report #592 : it was found on 2014 March, 18,
-// that WHERE was buggy due to OpenMP option
-// when n_elements of input bigger than !cpu.TPOOL_MIN_ELTS
-//
-// for WHERE, integers, also ptr and object
-template<class Sp>
-DLong* Data_<Sp>::Where( bool comp, SizeT& n)
-{
-  SizeT nEl = N_Elements();
-  DLong* ixList = new DLong[ nEl];
-  SizeT count = 0;
-  if( comp)
-    {
-      SizeT nIx = nEl;
-      //  TRACEOMP( __FILE__, __LINE__)
-      //#pragma omp parallel if (nEl >= CpuTPOOL_MIN_ELTS && (CpuTPOOL_MAX_ELTS == 0 || CpuTPOOL_MAX_ELTS <= nEl))
-    {
-      //#pragma omp for
-      for( SizeT i=0; i<nEl; ++i)
-	{
-	  if( (*this)[i] != 0)
-	    {
-	      ixList[ count++] = i;
-	    }
-	  else
-	    {
-	      ixList[ --nIx] = i;
-	    }
-	}
-    } // omp
-    }
-  else
-    //#pragma omp parallel if (nEl >= CpuTPOOL_MIN_ELTS && (CpuTPOOL_MAX_ELTS == 0 || CpuTPOOL_MAX_ELTS <= nEl))
-    {
-      //#pragma omp for
-    for( SizeT i=0; i<nEl; ++i)
-      if( (*this)[i] != 0)
-	{
-	  ixList[ count++] = i;
-	}
-    } // omp
-  n = count;
-  return ixList;
-}
-template<>
-DLong* Data_<SpDFloat>::Where( bool comp, SizeT& n)
-{
-  SizeT nEl = N_Elements();
-  DLong* ixList = new DLong[ nEl];
-  SizeT count = 0;
-  if( comp)
-    {
-    SizeT nIx = nEl;
-      //#pragma omp parallel if (nEl >= CpuTPOOL_MIN_ELTS && (CpuTPOOL_MAX_ELTS == 0 || CpuTPOOL_MAX_ELTS <= nEl))
-    {
-      //#pragma omp for
-      for( SizeT i=0; i<nEl; ++i)
-	{
-	  if( (*this)[i] != 0.0f)
-	    {
-          ixList[ count++] = i;
-        }
-	  else
-	    {
-          ixList[ --nIx] = i;
-        }
-      }
-    } // omp
-    }
-  else
-    //#pragma omp parallel if (nEl >= CpuTPOOL_MIN_ELTS && (CpuTPOOL_MAX_ELTS == 0 || CpuTPOOL_MAX_ELTS <= nEl))
-  {
-//#pragma omp for
-    for( SizeT i=0; i<nEl; ++i)
-      if( (*this)[i] != 0.0f)
-      {
-	  ixList[ count++] = i;
-      }
-    } // omp
-  n = count;
-  return ixList;
-  }
-template<>
-DLong* Data_<SpDDouble>::Where( bool comp, SizeT& n)
-{
-  SizeT nEl = N_Elements();
-  DLong* ixList = new DLong[ nEl];
-  SizeT count = 0;
-  if( comp)
-    {
-      SizeT nIx = nEl;
-      //#pragma omp parallel if (nEl >= CpuTPOOL_MIN_ELTS && (CpuTPOOL_MAX_ELTS == 0 || CpuTPOOL_MAX_ELTS <= nEl))
-    {
-      //#pragma omp for
-      for( SizeT i=0; i<nEl; ++i)
-	{
-	  if( (*this)[i] != 0.0)
-	    {
-	      ixList[ count++] = i;
-	    }
-	  else
-	    {
-	      ixList[ --nIx] = i;
-	    }
-	}
-    } // omp
-    }
-  else
-    //#pragma omp parallel if (nEl >= CpuTPOOL_MIN_ELTS && (CpuTPOOL_MAX_ELTS == 0 || CpuTPOOL_MAX_ELTS <= nEl))
-    {
-      //#pragma omp for
-    for( SizeT i=0; i<nEl; ++i)
-      if( (*this)[i] != 0.0)
-	{
-	  ixList[ count++] = i;
-	}
-    } // omp
-  n = count;
-  return ixList;
-}
-template<>
-DLong* Data_<SpDString>::Where( bool comp, SizeT& n)
-{
-  SizeT nEl = N_Elements();
-  DLong* ixList = new DLong[ nEl];
-  SizeT count = 0;
-  if( comp)
-    {
-      SizeT nIx = nEl;
-      //#pragma omp parallel if (nEl >= CpuTPOOL_MIN_ELTS && (CpuTPOOL_MAX_ELTS == 0 || CpuTPOOL_MAX_ELTS <= nEl))
-    {
-      //#pragma omp for
-      for( SizeT i=0; i<nEl; ++i)
-	{
-	  if( (*this)[i] != "")
-	    {
-	      ixList[ count++] = i;
-	    }
-	  else
-	    {
-	      ixList[ --nIx] = i;
-	    }
-	}
-    } // omp
-    }
-  else
-    //#pragma omp parallel if (nEl >= CpuTPOOL_MIN_ELTS && (CpuTPOOL_MAX_ELTS == 0 || CpuTPOOL_MAX_ELTS <= nEl))
-    {
-      //#pragma omp for
-    for( SizeT i=0; i<nEl; ++i)
-      if( (*this)[i] != "")
-	{
-	  ixList[ count++] = i;
-	}
-    } // omp
-  n = count;
-  return ixList;
-}
-template<>
-DLong* Data_<SpDComplex>::Where( bool comp, SizeT& n)
-{
-  SizeT nEl = N_Elements();
-  DLong* ixList = new DLong[ nEl];
-  SizeT count = 0;
-  if( comp)
-    {
-      SizeT nIx = nEl;
-      //#pragma omp parallel if (nEl >= CpuTPOOL_MIN_ELTS && (CpuTPOOL_MAX_ELTS == 0 || CpuTPOOL_MAX_ELTS <= nEl))
-    {
-      //#pragma omp for
-      for( SizeT i=0; i<nEl; ++i)
-	{
-	  if( (*this)[i].real() != 0.0 || (*this)[i].imag() != 0.0)
-	    {
-	      ixList[ count++] = i;
-	    }
-	  else
-	    {
-	      ixList[ --nIx] = i;
-	    }
-	}
-    } // omp
-    }
-  else
-    //#pragma omp parallel if (nEl >= CpuTPOOL_MIN_ELTS && (CpuTPOOL_MAX_ELTS == 0 || CpuTPOOL_MAX_ELTS <= nEl))
-    {
-      //#pragma omp for
-    for( SizeT i=0; i<nEl; ++i)
-      if( (*this)[i].real() != 0.0 || (*this)[i].imag() != 0.0)
-	{
-	  ixList[ count++] = i;
-	}
-    } // omp
-  n = count;
-  return ixList;
-}
-template<>
-DLong* Data_<SpDComplexDbl>::Where( bool comp, SizeT& n)
-{
-  SizeT nEl = N_Elements();
-  DLong* ixList = new DLong[ nEl];
-  SizeT count = 0;
-  if( comp)
-    {
-      SizeT nIx = nEl;
-      //#pragma omp parallel if (nEl >= CpuTPOOL_MIN_ELTS && (CpuTPOOL_MAX_ELTS == 0 || CpuTPOOL_MAX_ELTS <= nEl))
-    {
-      //#pragma omp for
-      for( SizeT i=0; i<nEl; ++i)
-	{
-	  if( (*this)[i].real() != 0.0 || (*this)[i].imag() != 0.0)
-	    {
-	      ixList[ count++] = i;
-	    }
-	  else
-	    {
-	      ixList[ --nIx] = i;
-	    }
-	}
-    } // omp
-    }
-  else
-    //#pragma omp parallel if (nEl >= CpuTPOOL_MIN_ELTS && (CpuTPOOL_MAX_ELTS == 0 || CpuTPOOL_MAX_ELTS <= nEl))
-    {
-      //#pragma omp for
-    for( SizeT i=0; i<nEl; ++i)
-      if( (*this)[i].real() != 0.0 || (*this)[i].imag() != 0.0)
-	{
-	  ixList[ count++] = i;
-	}
-    } // omp
-  n = count;
-  return ixList;
-}
-// structs are not allowed
-
-// SA: the start, stop, step & valIx arguments to MinMax are used when
-//     user specifies the DIMENSION keyword to MIN/MAX; they all have
-//     default values (defined in basegdl.hpp & datatypes.hpp) thus
-//     their introduction should not affect previous calls to MinMax
-
-// for use by MIN and MAX functions. 
-// New optimized (Parallel) version by GD.
-// integers: (NaN is not an issue) first
-template<class Sp>
-void Data_<Sp>::MinMax( DLong* minE, DLong* maxE, 
-			BaseGDL** minVal, BaseGDL** maxVal, bool omitNaN,
-            SizeT start, SizeT stop, SizeT step, DLong valIx, bool useAbs) {
-#define ABSFUNC llabs
-#include "minmax_include.cpp"
-#undef ABSFUNC
-}
-
-//string is identical, has no absfunc.
-template<>
-void Data_<SpDString>::MinMax( DLong* minE, DLong* maxE, 
-			       BaseGDL** minVal, BaseGDL** maxVal, bool omitNaN,
-                   SizeT start, SizeT stop, SizeT step, DLong valIx, bool useAbs)
-{
-#include "minmax_include.cpp"
-}
-
-
-// float etc. /nan is taken into account.
-// IDL treats +Inf and -Inf as normal (ordered) values unless /NAN is present. This complicates things.
-// Cases below are sped up by duplicating the code: NAN case and not.
-
-template<>
-void Data_<SpDFloat>::MinMax(DLong* minE, DLong* maxE,
-  BaseGDL** minVal, BaseGDL** maxVal, bool omitNaN,
-  SizeT start, SizeT stop, SizeT step, DLong valIx, bool useAbs) {
-#define ABSFUNC fabsf
-  if (omitNaN)
-  {
-#define MINMAX_HAS_OMITNAN
-#include "minmax_include.cpp"
-#undef MINMAX_HAS_OMITNAN
-  } else
-  {
-#include "minmax_include.cpp"
-  }
-#undef ABSFUNC
-}
-
-template<>
-void Data_<SpDDouble>::MinMax( DLong* minE, DLong* maxE, 
-			       BaseGDL** minVal, BaseGDL** maxVal, bool omitNaN,
-                   SizeT start, SizeT stop, SizeT step, DLong valIx, bool useAbs) {
-#define ABSFUNC fabs
-  if (omitNaN)
-  {
-#define MINMAX_HAS_OMITNAN
-#include "minmax_include.cpp"
-#undef MINMAX_HAS_OMITNAN
-  } else
-  {
-#include "minmax_include.cpp"
-  }
-#undef ABSFUNC
-}
-
-//special COMPLEX treatment, adds '.real()' where needed.
-#define MINMAX_IS_COMPLEX
-#define ABSFUNC std::abs
-template<>
-void Data_<SpDComplex>::MinMax( DLong* minE, DLong* maxE, 
-				BaseGDL** minVal, BaseGDL** maxVal, bool omitNaN, 
-                SizeT start, SizeT stop, SizeT step, DLong valIx, bool useAbs)
-{
-  if (omitNaN)
-  {
-#define MINMAX_HAS_OMITNAN
-#include "minmax_include.cpp"
-#undef MINMAX_HAS_OMITNAN
-  } else
-  {
-#include "minmax_include.cpp"
-  }
-}
-
-template<>
-void Data_<SpDComplexDbl>::MinMax( DLong* minE, DLong* maxE, 
-				   BaseGDL** minVal, BaseGDL** maxVal, bool omitNaN,
-                   SizeT start, SizeT stop, SizeT step, DLong valIx, bool useAbs)
-{
-  if (omitNaN)
-  {
-#define MINMAX_HAS_OMITNAN
-#include "minmax_include.cpp"
-#undef MINMAX_HAS_OMITNAN
-  } else
-  {
-#include "minmax_include.cpp"
-  }
-}
-#undef ABSFUNC
-#undef MINMAX_IS_COMPLEX
-
-template<>
-BaseGDL* Data_<SpDString>::Convol( BaseGDL* kIn, BaseGDL* scaleIn, BaseGDL* bias,
- 				   bool center, bool normalize, int edgeMode,
-                                   bool doNan, BaseGDL* missing, bool doMissing,
-                                   BaseGDL* invalid, bool doInvalid)
-{
-  throw GDLException("String expression not allowed in this context.");
-}
-template<>
-BaseGDL* Data_<SpDObj>::Convol( BaseGDL* kIn, BaseGDL* scaleIn, BaseGDL* bias,
- 				bool center, bool normalize, int edgeMode,
-                                bool doNan, BaseGDL* missing, bool doMissing,
-                                BaseGDL* invalid, bool doInvalid)
-{
-  throw GDLException("Object expression not allowed in this context.");
-}
-template<>
-BaseGDL* Data_<SpDPtr>::Convol( BaseGDL* kIn, BaseGDL* scaleIn,BaseGDL* bias,
- 				bool center, bool normalize, int edgeMode,
-                                bool doNan, BaseGDL* missing, bool doMissing,
-                                BaseGDL* invalid, bool doInvalid)
-{
-  throw GDLException("Pointer expression not allowed in this context.");
-}
-
-#define INCLUDE_CONVOL_CPP 1
-#define CONVOL_BYTE__
-
-#include "convol.cpp"
-
-#undef CONVOL_BYTE__
-#define CONVOL_UINT__
-
-#include "convol.cpp"
-
-#undef CONVOL_UINT__
-#define CONVOL_INT__
-
-#include "convol.cpp"
-
-#undef CONVOL_INT__
-#define CONVOL_ULONG__
-
-#include "convol.cpp"
-
-#undef CONVOL_ULONG__
-#define CONVOL_ULONG64__
-
-#include "convol.cpp"
-
-#undef CONVOL_ULONG64__
-
-#include "convol.cpp"
-
-//compute index when srcDim of rank rank is tranposed by [1,2,3...,0]. destStride must have been computed externally (accelerator).
-inline static SizeT transposed1Index(const SizeT inputindex, const SizeT* srcDim, const SizeT* destStride, const long rank)
-{
-  SizeT dim[MAXRANK];
-  SizeT index=inputindex;
-  SizeT sizeleft = index;
-  SizeT dim0;
-  for (SizeT i = 0; i < rank; ++i) {
-    dim0= srcDim[i];
-    sizeleft /= dim0;
-    dim[i] = index - sizeleft * dim0; //the corresponding index in src data.
-    index = sizeleft;
-  }
-  SizeT ix = 0;
-  for (SizeT i = 0; i < rank-1; ++i) ix += dim[i+1] * destStride[i]; ix+=dim[0]*destStride[rank-1]; //note dim has been rotated by 1
-
-  return ix;
-}
-#define INCLUDE_SMOOTH_POLYD
-template<typename T>
-void SmoothPolyD(T* srcIn, T* destIn, const SizeT* datainDim, const int rank, const DLong* width) {
- #include "smoothPolyD.hpp" 
-}
-//subset having edges
-#define USE_EDGE
-template<typename T>
-void SmoothPolyDWrap(T* srcIn, T* destIn, const SizeT* datainDim, const int rank, const DLong* width) {
-#define EDGE_WRAP
-#include "smoothPolyD.hpp" 
-#undef EDGE_WRAP
-}
-template<typename T>
-void SmoothPolyDTruncate(T* srcIn, T* destIn, const SizeT* datainDim, const int rank, const DLong* width) {
-#define EDGE_TRUNCATE
-#include "smoothPolyD.hpp" 
-#undef EDGE_TRUNCATE
-}
-template<typename T>
-void SmoothPolyDZero(T* srcIn, T* destIn, const SizeT* datainDim, const int rank, const DLong* width) {
-#define EDGE_ZERO
-#include "smoothPolyD.hpp" 
-#undef EDGE_ZERO
-}
-template<typename T>
-void SmoothPolyDMirror(T* srcIn, T* destIn, const SizeT* datainDim, const int rank, const DLong* width) {
-#define EDGE_MIRROR
-#include "smoothPolyD.hpp" 
-#undef EDGE_MIRROR
-}
-#undef USE_EDGE
-#undef INCLUDE_SMOOTH_POLYD
-
-#define INCLUDE_SMOOTH_POLYD_NAN
-template<typename T>
-void SmoothPolyDNan(T* srcIn, T* destIn, const SizeT* datainDim, const int rank, const DLong* width) {
- #include "smoothPolyDnans.hpp" 
-}
-//subset having edges
-#define USE_EDGE
-template<typename T>
-void SmoothPolyDWrapNan(T* srcIn, T* destIn, const SizeT* datainDim, const int rank, const DLong* width) {
-#define EDGE_WRAP
-#include "smoothPolyDnans.hpp" 
-#undef EDGE_WRAP
-}
-template<typename T>
-void SmoothPolyDTruncateNan(T* srcIn, T* destIn, const SizeT* datainDim, const int rank, const DLong* width) {
-#define EDGE_TRUNCATE
-#include "smoothPolyDnans.hpp" 
-#undef EDGE_TRUNCATE
-}
-template<typename T>
-void SmoothPolyDZeroNan(T* srcIn, T* destIn, const SizeT* datainDim, const int rank, const DLong* width) {
-#define EDGE_ZERO
-#include "smoothPolyDnans.hpp" 
-#undef EDGE_ZERO
-}
-template<typename T>
-void SmoothPolyDMirrorNan(T* srcIn, T* destIn, const SizeT* datainDim, const int rank, const DLong* width) {
-#define EDGE_MIRROR
-#include "smoothPolyDnans.hpp" 
-#undef EDGE_MIRROR
-}
-#undef USE_EDGE
-#undef INCLUDE_SMOOTH_POLYD_NAN
-
-
-#define INCLUDE_SMOOTH_1D
-template<typename T>
-void Smooth1D(T* data, T* res, SizeT dimx, SizeT w) {
-#include "smooth1d.hpp"
-}
-//subset having edges
-#define USE_EDGE
-template<typename T>
-void Smooth1DWrap(T* data, T* res, SizeT dimx, SizeT w) {
-#define EDGE_WRAP
-#include "smooth1d.hpp"
-#undef EDGE_WRAP
-}
-template<typename T>
-void Smooth1DTruncate(T* data, T* res, SizeT dimx, SizeT w) {
-#define EDGE_TRUNCATE
-#include "smooth1d.hpp"
-#undef EDGE_TRUNCATE
-}
-template<typename T>
-void Smooth1DMirror(T* data, T* res, SizeT dimx, SizeT w) {
-#define EDGE_MIRROR
-#include "smooth1d.hpp"
-#undef EDGE_MIRROR
-}
-template<typename T>
-void Smooth1DZero(T* data, T* res, SizeT dimx, SizeT w) {
-#define EDGE_ZERO
-#include "smooth1d.hpp"
-#undef EDGE_ZERO
-}
-#undef USE_EDGE
-#undef INCLUDE_SMOOTH_1D
-
-//smooth 1d functions for Nans.
-#define INCLUDE_SMOOTH_1D_NAN
-template<typename T>
-void Smooth1DNan(T* data, T* res, SizeT dimx, SizeT w) {
-#include "smooth1dnans.hpp"
-}
-//subset having edges
-#define USE_EDGE
-template<typename T>
-void Smooth1DWrapNan(T* data, T* res, SizeT dimx, SizeT w) {
-#define EDGE_WRAP
-#include "smooth1dnans.hpp"
-#undef EDGE_WRAP
-}
-template<typename T>
-void Smooth1DTruncateNan(T* data, T* res, SizeT dimx, SizeT w) {
-#define EDGE_TRUNCATE
-#include "smooth1dnans.hpp"
-#undef EDGE_TRUNCATE
-}
-template<typename T>
-void Smooth1DMirrorNan(T* data, T* res, SizeT dimx, SizeT w) {
-#define EDGE_MIRROR
-#include "smooth1dnans.hpp"
-#undef EDGE_MIRROR
-}
-template<typename T>
-void Smooth1DZeroNan(T* data, T* res, SizeT dimx, SizeT w) {
-#define EDGE_ZERO
-#include "smooth1dnans.hpp"
-#undef EDGE_ZERO
-}
-#undef USE_EDGE
-#undef INCLUDE_SMOOTH_1D_NAN
-
-//smooth 2d functions.
-#define INCLUDE_SMOOTH_2D
-template<typename T>
-void Smooth2D(const T* src, T* dest, const SizeT dimx, const SizeT dimy, const DLong* width) {
-#include "smooth2d.hpp"
-}
-//subset having edges
-#define USE_EDGE
-template<typename T>
-void Smooth2DWrap(const T* src, T* dest, const SizeT dimx, const SizeT dimy, const DLong* width) {
-#define EDGE_WRAP
-#include "smooth2d.hpp"
-#undef EDGE_WRAP
-}
-template<typename T>
-void Smooth2DTruncate(const T* src, T* dest, const SizeT dimx, const SizeT dimy, const DLong* width) {
-#define EDGE_TRUNCATE
-#include "smooth2d.hpp"
-#undef EDGE_TRUNCATE
-}
-template<typename T>
-void Smooth2DMirror(const T* src, T* dest, const SizeT dimx, const SizeT dimy, const DLong* width) {
-#define EDGE_MIRROR
-#include "smooth2d.hpp"
-#undef EDGE_MIRROR
-}
-template<typename T>
-void Smooth2DZero(const T* src, T* dest, const SizeT dimx, const SizeT dimy, const DLong* width) {
-#define EDGE_ZERO
-#include "smooth2d.hpp"
-#undef EDGE_ZERO
-}
-#undef USE_EDGE
-#undef INCLUDE_SMOOTH_2D
-
-//smooth 2d functions for Nans.
-#define INCLUDE_SMOOTH_2D_NAN
-template<typename T>
-void Smooth2DNan(const T* src, T* dest, const SizeT dimx, const SizeT dimy, const DLong* width) {
-#include "smooth2dnans.hpp"
-}
-//subset having edges
-#define USE_EDGE
-template<typename T>
-void Smooth2DWrapNan(const T* src, T* dest, const SizeT dimx, const SizeT dimy, const DLong* width) {
-#define EDGE_WRAP
-#include "smooth2dnans.hpp"
-#undef EDGE_WRAP
-}
-template<typename T>
-void Smooth2DTruncateNan(const T* src, T* dest, const SizeT dimx, const SizeT dimy, const DLong* width) {
-#define EDGE_TRUNCATE
-#include "smooth2dnans.hpp"
-#undef EDGE_TRUNCATE
-}
-template<typename T>
-void Smooth2DMirrorNan(const T* src, T* dest, const SizeT dimx, const SizeT dimy, const DLong* width) {
-#define EDGE_MIRROR
-#include "smooth2dnans.hpp"
-#undef EDGE_MIRROR
-}
-template<typename T>
-void Smooth2DZeroNan(const T* src, T* dest, const SizeT dimx, const SizeT dimy, const DLong* width) {
-#define EDGE_ZERO
-#include "smooth2dnans.hpp"
-#undef EDGE_ZERO
-}
-#undef USE_EDGE
-#undef INCLUDE_SMOOTH_2D_NAN
-
-//Note: Values for ULong types return differently as IDL, but it should be proven that IDL is right...
-template<class Sp>
-BaseGDL* Data_<Sp>::Smooth(DLong* width, int edgeMode,
-                                bool doNan, BaseGDL* missing)
-{
-  Ty missingValue = (*static_cast<Data_*>( missing))[0];
-  SizeT nA = N_Elements();
-  
-  SizeT srcRank = this->Rank();
-  
-  BaseGDL* data = this;
-  BaseGDL* res = this->Dup();
-  
-  SizeT sum = 0;
-  SizeT mydims[srcRank]; //memory of dimensions, done one after the other
-  for (int i = 0; i < srcRank; ++i) 
-  {
-    mydims[i] = this->Dim(i);
-    sum += width[i];
-  }
-  if (sum == srcRank) return res;
-  
-  DUInt rotate1[srcRank]; //turntable transposition index list
-  
-  //doNan only meaningful for real and double (complex are real and double here)
-  doNan=(doNan && (this->Type()==GDL_FLOAT ||this->Type()==GDL_DOUBLE));
-
-  for (int i = 0; i < srcRank-1; ++i) rotate1[i] = i+1; rotate1[srcRank-1]=0; //[1,2,...,0] 
- 
-  if (doNan) {
-    if (srcRank==1) {
-      SizeT dimx = nA;
-      SizeT w = width[0] / 2;
-      if (edgeMode==0){
-        Smooth1DNan((Ty*)data->DataAddr(), (Ty*)res->DataAddr(),  dimx, w);
-      } else if (edgeMode==1) {
-        Smooth1DWrapNan((Ty*)data->DataAddr(), (Ty*)res->DataAddr(),  dimx, w);
-      } else if (edgeMode==2) {
-        Smooth1DTruncateNan((Ty*)data->DataAddr(), (Ty*)res->DataAddr(),  dimx, w);
-      } else if (edgeMode==3) {
-        Smooth1DZeroNan((Ty*)data->DataAddr(), (Ty*)res->DataAddr(),  dimx, w);
-      } else if (edgeMode==4) {
-        Smooth1DMirrorNan((Ty*)data->DataAddr(), (Ty*)res->DataAddr(),  dimx, w);
-      }
-    } else if (srcRank==2) {
-      SizeT dimx = data->Dim(0);
-      SizeT dimy = data->Dim(1);
-      if (edgeMode==0){
-        Smooth2DNan((Ty*)data->DataAddr(), (Ty*)res->DataAddr(), dimx, dimy,  width);
-      } else if (edgeMode==1) {
-        Smooth2DWrapNan((Ty*)data->DataAddr(), (Ty*)res->DataAddr(), dimx, dimy,  width);
-      } else if (edgeMode==2) {
-        Smooth2DTruncateNan((Ty*)data->DataAddr(), (Ty*)res->DataAddr(), dimx, dimy,  width);
-      } else if (edgeMode==3) {
-        Smooth2DZeroNan((Ty*)data->DataAddr(), (Ty*)res->DataAddr(), dimx, dimy,  width);
-      } else if (edgeMode==4) {
-        Smooth2DMirrorNan((Ty*)data->DataAddr(), (Ty*)res->DataAddr(), dimx, dimy,  width);
-      }
-    } else {
-      long rank = data->Rank();
-      SizeT srcDim[MAXRANK];
-      for (int i = 0; i < rank; ++i) srcDim[i] = data->Dim()[i];
-      BaseGDL* dataw = this->Dup();
-      if (edgeMode==0){
-        SmoothPolyDNan((Ty*)dataw->DataAddr(), (Ty*)res->DataAddr(), srcDim, rank, width);
-      } else if (edgeMode==1) {
-        SmoothPolyDWrapNan((Ty*)dataw->DataAddr(), (Ty*)res->DataAddr(), srcDim, rank, width);
-      } else if (edgeMode==2) {
-        SmoothPolyDTruncateNan((Ty*)dataw->DataAddr(), (Ty*)res->DataAddr(), srcDim, rank, width);
-      } else if (edgeMode==3) {
-        SmoothPolyDZeroNan((Ty*)dataw->DataAddr(), (Ty*)res->DataAddr(), srcDim, rank, width);
-      } else if (edgeMode==4) {
-        SmoothPolyDMirrorNan((Ty*)dataw->DataAddr(), (Ty*)res->DataAddr(), srcDim, rank, width);
-      }
-      GDLDelete(dataw);
-    }
-   //additional loop to replace left Nans by missing, if nans are left anyway
-    if (gdlValid(missingValue))  {
-      Ty* resty=(Ty*)res->DataAddr();
-#pragma omp for 
-      for (SizeT i=0; i<nA; ++i) if (!gdlValid(resty[i])) resty[i]=missingValue;
-      }
-  } else {
-    if (srcRank==1) {
-      SizeT dimx = nA;
-      SizeT w = width[0] / 2;
-      if (edgeMode==0){
-        Smooth1D((Ty*)data->DataAddr(), (Ty*)res->DataAddr(),  dimx, w);
-      } else if (edgeMode==1) {
-        Smooth1DWrap((Ty*)data->DataAddr(), (Ty*)res->DataAddr(),  dimx, w);
-      } else if (edgeMode==2) {
-        Smooth1DTruncate((Ty*)data->DataAddr(), (Ty*)res->DataAddr(),  dimx, w);
-      } else if (edgeMode==3) {
-        Smooth1DZero((Ty*)data->DataAddr(), (Ty*)res->DataAddr(),  dimx, w);
-      } else if (edgeMode==4) {
-        Smooth1DMirror((Ty*)data->DataAddr(), (Ty*)res->DataAddr(),  dimx, w);
-      }
-    } else if (srcRank==2) {
-      SizeT dimx = data->Dim(0);
-      SizeT dimy = data->Dim(1);
-      if (edgeMode==0){
-        Smooth2D((Ty*)data->DataAddr(), (Ty*)res->DataAddr(), dimx, dimy,  width);
-      } else if (edgeMode==1) {
-        Smooth2DWrap((Ty*)data->DataAddr(), (Ty*)res->DataAddr(), dimx, dimy,  width);
-      } else if (edgeMode==2) {
-        Smooth2DTruncate((Ty*)data->DataAddr(), (Ty*)res->DataAddr(), dimx, dimy,  width);
-      } else if (edgeMode==3) {
-        Smooth2DZero((Ty*)data->DataAddr(), (Ty*)res->DataAddr(), dimx, dimy,  width);
-      } else if (edgeMode==4) {
-        Smooth2DMirror((Ty*)data->DataAddr(), (Ty*)res->DataAddr(), dimx, dimy,  width);
-      }
-    } else  {
-      long rank = data->Rank();
-      SizeT srcDim[MAXRANK];
-      for (int i = 0; i < rank; ++i) srcDim[i] = data->Dim()[i];
-      BaseGDL* dataw = this->Dup();
-      if (edgeMode==0){
-        SmoothPolyD((Ty*)dataw->DataAddr(), (Ty*)res->DataAddr(), srcDim, rank, width);
-      } else if (edgeMode==1) {
-        SmoothPolyDWrap((Ty*)dataw->DataAddr(), (Ty*)res->DataAddr(), srcDim, rank, width);
-      } else if (edgeMode==2) {
-        SmoothPolyDTruncate((Ty*)dataw->DataAddr(), (Ty*)res->DataAddr(), srcDim, rank, width);
-      } else if (edgeMode==3) {
-        SmoothPolyDZero((Ty*)dataw->DataAddr(), (Ty*)res->DataAddr(), srcDim, rank, width);
-      } else if (edgeMode==4) {
-        SmoothPolyDMirror((Ty*)dataw->DataAddr(), (Ty*)res->DataAddr(), srcDim, rank, width);
-      }
-      GDLDelete(dataw);
-    }
-  }
-  return res;
-}
-
-template<>
-BaseGDL* Data_<SpDString>::Smooth( DLong* width, int edgeMode,
-                                   bool doNan, BaseGDL* missing)
-{
-  throw GDLException("String expression not allowed in this context.");
-}
-template<>
-BaseGDL* Data_<SpDObj>::Smooth( DLong* width, int edgeMode,
-                                bool doNan, BaseGDL* missing)
-{
-  throw GDLException("Object expression not allowed in this context.");
-}
-template<>
-BaseGDL* Data_<SpDPtr>::Smooth( DLong* width, int edgeMode,
-                                bool doNan, BaseGDL* missing)
-{
-  throw GDLException("Pointer expression not allowed in this context.");
-}
-template<>
-BaseGDL* Data_<SpDComplexDbl>::Smooth( DLong* width, int edgeMode,
-                                bool doNan, BaseGDL* missing)
-{
-  Ty missingValue = (*static_cast<Data_*>( missing))[0];
-  DDoubleGDL* missr=new DDoubleGDL(missingValue.real());
-  DDoubleGDL* missi=new DDoubleGDL(missingValue.imag());
-  Data_* res = this->Dup();
-  DDoubleGDL* re=new DDoubleGDL(this->Dim(), BaseGDL::NOZERO);
-  for (SizeT i=0; i< this->N_Elements(); ++i) (*re)[i]=(*this)[i].real();
-  BaseGDL* resr=re->Smooth(width, edgeMode, doNan, missr);
-  DDoubleGDL* im=new DDoubleGDL(this->Dim(), BaseGDL::NOZERO);
-  for (SizeT i=0; i< this->N_Elements(); ++i) (*im)[i]=(*this)[i].imag();
-  BaseGDL* resi=im->Smooth(width, edgeMode, doNan, missi);
-  DDouble* dresi=(DDouble*)resi->DataAddr();  
-  DDouble* dresr=(DDouble*)resr->DataAddr();  
-  for (SizeT i=0; i< this->N_Elements(); ++i) (*res)[i]=std::complex<DDouble>(dresr[i], dresr[i]);
-  GDLDelete (resr);
-  GDLDelete (re);
-  GDLDelete (missr);
-  GDLDelete (resi);
-  GDLDelete (im);
-  GDLDelete (missi);
-  return res;
-}
-template<>
-BaseGDL* Data_<SpDComplex>::Smooth( DLong* width, int edgeMode,
-                                bool doNan, BaseGDL* missing)
-{
-  Ty missingValue = (*static_cast<Data_*>( missing))[0];
-  DFloatGDL* missr=new DFloatGDL(missingValue.real());
-  DFloatGDL* missi=new DFloatGDL(missingValue.imag());
-  Data_* res = this->Dup();
-  DFloatGDL* re=new DFloatGDL(this->Dim(), BaseGDL::NOZERO);
-  for (SizeT i=0; i< this->N_Elements(); ++i) (*re)[i]=(*this)[i].real();
-  BaseGDL* resr=re->Smooth(width, edgeMode, doNan, missr);
-  DFloatGDL* im=new DFloatGDL(this->Dim(), BaseGDL::NOZERO);
-  for (SizeT i=0; i< this->N_Elements(); ++i) (*im)[i]=(*this)[i].imag();
-  BaseGDL* resi=im->Smooth(width, edgeMode, doNan, missi);
-  DFloat* fresi=(DFloat*)resi->DataAddr();  
-  DFloat* fresr=(DFloat*)resr->DataAddr();  
-  for (SizeT i=0; i< this->N_Elements(); ++i) (*res)[i]=std::complex<DFloat>(fresr[i], fresr[i]);
-  GDLDelete (resr);
-  GDLDelete (re);
-  GDLDelete (missr);
-  GDLDelete (resi);
-  GDLDelete (im);
-  GDLDelete (missi);
-  return res;
-}
 
 template<>
 BaseGDL* Data_<SpDString>::Rebin( const dimension& newDim, bool sample)
