+++ src/smooth.cpp
@@ -3,6 +3,7 @@
                              -------------------
     begin                : 05 September 2014
     copyright            : (C) 2014 by Levan Loria  (with Alain Coulais)
+                         : (C) 2017 by Gilles Duvert 
     email                : alaingdl@users.sourceforge.net
  ***************************************************************************/
 
@@ -16,6 +17,187 @@
  ***************************************************************************/
 
 #include "includefirst.hpp"
+
+#if defined(USE_PYTHON) || defined(PYTHON_MODULE)
+#include <numpy/arrayobject.h>
+#endif
+
+#include "nullgdl.hpp"
+#include "dstructgdl.hpp"
+#include "dinterpreter.hpp"
+#ifdef _MSC_VER
+#define isfinite _finite
+#define std__isnan isnan
+#else
+#define std__isnan std::isnan
+#endif
+
+using std::isfinite;
+
+template<typename T>
+inline bool gdlValid( const T &value )
+{
+    T max_value = std::numeric_limits<T>::max();
+    T min_value = - max_value;
+    return ( ( min_value <= value && value <= max_value ) &&  (value == value));
+}
+inline bool gdlValid( const DComplex &value )
+{
+    DFloat max_value = std::numeric_limits<DFloat>::max();
+    DFloat min_value = - max_value;
+    return ( ( min_value <= value.real() && value.real() <= max_value ) &&  (value.real() == value.real()))&&
+            ( ( min_value <= value.imag() && value.imag() <= max_value ) && (value.imag() == value.imag()));
+}
+inline bool gdlValid( const DComplexDbl &value )
+{
+    DDouble max_value = std::numeric_limits<DDouble>::max();
+    DDouble min_value = - max_value;
+    return ( ( min_value <= value.real() && value.real() <= max_value ) &&  (value.real() == value.real()))&&
+            ( ( min_value <= value.imag() && value.imag() <= max_value ) &&  (value.imag() == value.imag()));
+}
+
+//compute index when srcDim of rank rank is tranposed by [1,2,3...,0]. destStride must have been computed externally (accelerator).
+inline static SizeT transposed1Index(const SizeT inputindex, const SizeT* srcDim, const SizeT* destStride, const long rank)
+{
+  SizeT dim[MAXRANK];
+  SizeT index=inputindex;
+  SizeT sizeleft = index;
+  SizeT dim0;
+  for (SizeT i = 0; i < rank; ++i) {
+    dim0= srcDim[i];
+    sizeleft /= dim0;
+    dim[i] = index - sizeleft * dim0; //the corresponding index in src data.
+    index = sizeleft;
+  }
+  SizeT ix = 0;
+  for (SizeT i = 0; i < rank-1; ++i) ix += dim[i+1] * destStride[i]; ix+=dim[0]*destStride[rank-1]; //note dim has been rotated by 1
+
+  return ix;
+}
+// include repeatedly smooth_inc for all useful types.
+
+#define SMOOTH_Ty DByte
+#define SMOOTH_SP SpDByte
+#include "smooth_inc.cpp"
+#undef SMOOTH_Ty
+#undef SMOOTH_SP
+
+#define SMOOTH_Ty DInt
+#define SMOOTH_SP SpDInt
+#include "smooth_inc.cpp"
+#undef SMOOTH_Ty
+#undef SMOOTH_SP
+
+#define SMOOTH_Ty DUInt
+#define SMOOTH_SP SpDUInt
+#include "smooth_inc.cpp"
+#undef SMOOTH_Ty
+#undef SMOOTH_SP
+
+#define SMOOTH_Ty DLong
+#define SMOOTH_SP SpDLong
+#include "smooth_inc.cpp"
+#undef SMOOTH_Ty
+#undef SMOOTH_SP
+
+#define SMOOTH_Ty DULong
+#define SMOOTH_SP SpDULong
+#include "smooth_inc.cpp"
+#undef SMOOTH_Ty
+#undef SMOOTH_SP
+
+#define SMOOTH_Ty DLong64
+#define SMOOTH_SP SpDLong64
+#include "smooth_inc.cpp"
+#undef SMOOTH_Ty
+#undef SMOOTH_SP
+
+#define SMOOTH_Ty DULong64
+#define SMOOTH_SP SpDULong64
+#include "smooth_inc.cpp"
+#undef SMOOTH_Ty
+#undef SMOOTH_SP
+
+#define SMOOTH_Ty DFloat
+#define SMOOTH_SP SpDFloat
+#include "smooth_inc.cpp"
+#undef SMOOTH_Ty
+#undef SMOOTH_SP
+
+#define SMOOTH_Ty DDouble
+#define SMOOTH_SP SpDDouble
+#include "smooth_inc.cpp"
+#undef SMOOTH_Ty
+#undef SMOOTH_SP
+
+template<>
+BaseGDL* Data_<SpDString>::Smooth( DLong* width, int edgeMode,
+                                   bool doNan, BaseGDL* missing)
+{
+  throw GDLException("String expression not allowed in this context.");
+}
+template<>
+BaseGDL* Data_<SpDObj>::Smooth( DLong* width, int edgeMode,
+                                bool doNan, BaseGDL* missing)
+{
+  throw GDLException("Object expression not allowed in this context.");
+}
+template<>
+BaseGDL* Data_<SpDPtr>::Smooth( DLong* width, int edgeMode,
+                                bool doNan, BaseGDL* missing)
+{
+  throw GDLException("Pointer expression not allowed in this context.");
+}
+template<>
+BaseGDL* Data_<SpDComplexDbl>::Smooth( DLong* width, int edgeMode,
+                                bool doNan, BaseGDL* missing)
+{
+  Ty missingValue = (*static_cast<Data_*>( missing))[0];
+  DDoubleGDL* missr=new DDoubleGDL(missingValue.real());
+  DDoubleGDL* missi=new DDoubleGDL(missingValue.imag());
+  Data_* res = this->Dup();
+  DDoubleGDL* re=new DDoubleGDL(this->Dim(), BaseGDL::NOZERO);
+  for (SizeT i=0; i< this->N_Elements(); ++i) (*re)[i]=(*this)[i].real();
+  BaseGDL* resr=re->Smooth(width, edgeMode, doNan, missr);
+  DDoubleGDL* im=new DDoubleGDL(this->Dim(), BaseGDL::NOZERO);
+  for (SizeT i=0; i< this->N_Elements(); ++i) (*im)[i]=(*this)[i].imag();
+  BaseGDL* resi=im->Smooth(width, edgeMode, doNan, missi);
+  DDouble* dresi=(DDouble*)resi->DataAddr();  
+  DDouble* dresr=(DDouble*)resr->DataAddr();  
+  for (SizeT i=0; i< this->N_Elements(); ++i) (*res)[i]=std::complex<DDouble>(dresr[i], dresr[i]);
+  GDLDelete (resr);
+  GDLDelete (re);
+  GDLDelete (missr);
+  GDLDelete (resi);
+  GDLDelete (im);
+  GDLDelete (missi);
+  return res;
+}
+template<>
+BaseGDL* Data_<SpDComplex>::Smooth( DLong* width, int edgeMode,
+                                bool doNan, BaseGDL* missing)
+{
+  Ty missingValue = (*static_cast<Data_*>( missing))[0];
+  DFloatGDL* missr=new DFloatGDL(missingValue.real());
+  DFloatGDL* missi=new DFloatGDL(missingValue.imag());
+  Data_* res = this->Dup();
+  DFloatGDL* re=new DFloatGDL(this->Dim(), BaseGDL::NOZERO);
+  for (SizeT i=0; i< this->N_Elements(); ++i) (*re)[i]=(*this)[i].real();
+  BaseGDL* resr=re->Smooth(width, edgeMode, doNan, missr);
+  DFloatGDL* im=new DFloatGDL(this->Dim(), BaseGDL::NOZERO);
+  for (SizeT i=0; i< this->N_Elements(); ++i) (*im)[i]=(*this)[i].imag();
+  BaseGDL* resi=im->Smooth(width, edgeMode, doNan, missi);
+  DFloat* fresi=(DFloat*)resi->DataAddr();  
+  DFloat* fresr=(DFloat*)resr->DataAddr();  
+  for (SizeT i=0; i< this->N_Elements(); ++i) (*res)[i]=std::complex<DFloat>(fresr[i], fresr[i]);
+  GDLDelete (resr);
+  GDLDelete (re);
+  GDLDelete (missr);
+  GDLDelete (resi);
+  GDLDelete (im);
+  GDLDelete (missi);
+  return res;
+}
 
 #include "datatypes.hpp"
 #include "envt.hpp"
