+++ src/tagwhere_inc.cpp
@@ -1,9 +1,9 @@
 /***************************************************************************
-                          convol.hpp  -  Convolution GDL library function
+                          tagwhere_inc.cpp  -  include for where.cpp
                              -------------------
-    begin                : Jun 21 2013
-    copyright            : (C) 2013, G. Duvert 2017
-    email                : 
+    begin                : Apr 7 2018
+    copyright            : (C) 2002 by Marc Schellens, G. Duvert
+    email                : m_schellens@users.sf.net
  ***************************************************************************/
 
 /***************************************************************************
@@ -15,22 +15,21 @@
  *                                                                         *
  ***************************************************************************/
 
- #ifdef HAVE_CONFIG_H
- #include <config.h>
- #endif
-
-
-#ifndef CONVOL_HPP_
-#define CONVOL_HPP_
-
-#include "datatypes.hpp"
-#include "envt.hpp"
-
-namespace lib {
-
-  BaseGDL* convol_fun( EnvT* e);
-
-} // namespace
-
-
-#endif
+template<>
+DByte* Data_<Sp>::TagWhere(SizeT& n)
+{
+  SizeT nEl = N_Elements();
+  DByte* ixList = new DByte[ nEl];
+  SizeT count = 0;
+#pragma omp parallel reduction(+:count) if (nEl >= CpuTPOOL_MIN_ELTS && (CpuTPOOL_MAX_ELTS == 0 || CpuTPOOL_MAX_ELTS <= nEl))
+    {
+#pragma omp for 
+    for (SizeT i = 0; i < nEl; ++i) {
+      DByte tmp=(dd[i]!=0);
+      ixList[i]=tmp;
+      count +=tmp;
+    }
+    }
+  n = count;
+  return ixList;    
+}
